requirements.txt
kafka-python
numpy
pandas
plotly
requests
river
scikit-learn
streamlit
streamlit-autorefresh
yfinance